Revert "Merge pull request #106 from LLRHook/feature/hybrid-ocr-compare-view"
This reverts commit 871c6d030d3124d7d028c6f268486101b201420c, reversing
changes made to cc6d58c4cdefad0d885680358818c6ee359c9dab.
