+++ backend/internal/receipt/service_test.go
@@ -260,74 +260,6 @@ func TestParseResponseText_RestaurantReceipt(t *testing.T) {
 	}
 }
 
-func TestParseResponseText_WithBoundingBoxes(t *testing.T) {
-	input := `{
-		"vendor": "Kroger",
-		"items": [
-			{"name": "Milk", "price": 3.99, "bounding_box": {"x": 0.05, "y": 0.20, "width": 0.90, "height": 0.04}},
-			{"name": "Bread", "price": 2.49}
-		],
-		"total": 6.48
-	}`
-
-	receipt, err := parseResponseText(input)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if len(receipt.Items) != 2 {
-		t.Fatalf("items count = %d, want 2", len(receipt.Items))
-	}
-
-	// First item should have bounding box
-	if receipt.Items[0].BoundingBox == nil {
-		t.Fatal("items[0].bounding_box should not be nil")
-	}
-	if receipt.Items[0].BoundingBox.X != 0.05 {
-		t.Errorf("items[0].bounding_box.x = %f, want 0.05", receipt.Items[0].BoundingBox.X)
-	}
-	if receipt.Items[0].BoundingBox.Y != 0.20 {
-		t.Errorf("items[0].bounding_box.y = %f, want 0.20", receipt.Items[0].BoundingBox.Y)
-	}
-	if receipt.Items[0].BoundingBox.Width != 0.90 {
-		t.Errorf("items[0].bounding_box.width = %f, want 0.90", receipt.Items[0].BoundingBox.Width)
-	}
-	if receipt.Items[0].BoundingBox.Height != 0.04 {
-		t.Errorf("items[0].bounding_box.height = %f, want 0.04", receipt.Items[0].BoundingBox.Height)
-	}
-
-	// Second item should have no bounding box
-	if receipt.Items[1].BoundingBox != nil {
-		t.Errorf("items[1].bounding_box should be nil, got %+v", receipt.Items[1].BoundingBox)
-	}
-}
-
-func TestParseResponseText_WithRawOcrName(t *testing.T) {
-	input := `{
-		"vendor": "Walmart",
-		"items": [
-			{"name": "Chicken Breast", "price": 8.99, "raw_ocr_name": "CHKN BRST"},
-			{"name": "Milk", "price": 3.99}
-		],
-		"total": 12.98
-	}`
-
-	receipt, err := parseResponseText(input)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if len(receipt.Items) != 2 {
-		t.Fatalf("items count = %d, want 2", len(receipt.Items))
-	}
-	if receipt.Items[0].RawOcrName != "CHKN BRST" {
-		t.Errorf("items[0].raw_ocr_name = %q, want %q", receipt.Items[0].RawOcrName, "CHKN BRST")
-	}
-	if receipt.Items[1].RawOcrName != "" {
-		t.Errorf("items[1].raw_ocr_name = %q, want empty string", receipt.Items[1].RawOcrName)
-	}
-}
-
 func TestTruncate(t *testing.T) {
 	if got := truncate("hello", 10); got != "hello" {
 		t.Errorf("truncate short = %q, want %q", got, "hello")
