+++ mobile/ios/Flutter/AppFrameworkInfo.plist
@@ -20,5 +20,7 @@
   <string>????</string>
   <key>CFBundleVersion</key>
   <string>1.0</string>
+  <key>MinimumOSVersion</key>
+  <string>13.0</string>
 </dict>
 </plist>
