+++ mobile/pubspec.yaml
@@ -22,7 +22,6 @@ dependencies:
   http: ^1.5.0
   logger: ^2.6.2
   image_picker: ^1.1.2
-  google_mlkit_text_recognition: ^0.15.1
 
 dev_dependencies:
   flutter_test:
